Replace Sandbox @timeout with @timeout_data for udata
The goal is to alter the behavior of Sandbox so that the execution
timeout is measured from the time of calling #eval, instead of the time
of creating the sandbox.

To achieve this, instead of placing a timeout struct on the heap and
passing that as the user data pointer, we allocate a mutable Crystal
class with a start point field that we can update prior to engaging the
VM.

I also had to do two monkeypatches to the Context methods that serve as
the edge of Crystal code before calling into LibDuk. This may not be the
proper way to do this, but it is an awkward situation, and Sandbox is
the only descendant of Context, so I'm not sure what the intention is..

The next commit will complete the wiring of this new object.

---

As a bonus, this adds a ctor that takes a time span instead of just an
int, so you can expressively decalre the timeout as such:

    Duktape::Sandbox.new(500.milliseoncds)

instead of having to check what units are expected.