Replace duk_cr_timeout with Crystal function
It is not well known, but Crystal allows you to export symbols that can
be called by foreign code. As such, this replaces the C procedure
definted in the duk_config.h header with a Crystal native `fun`
definition.

In this `fun`, we cast down the user data pointer - a pointer to our
TimeoutData struct for the executing Sandbox instance - and the timeout
measurement is made.

This could have been done by altering the C program as well. But I
figured that moving this to native Crystal code would be preferred.